requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
openai
python-dotenv
